fixed fast refresh error by separating context hooks & providers
