+++ frontend/src/providers/ModeContext.tsx
@@ -1,13 +1,5 @@
-import React, { createContext, useContext, useState } from "react";
-
-export type Mode = "student" | "instructor";
-
-type ModeContextValue = {
-  mode: Mode;
-  setMode: (m: Mode) => void;
-};
-
-const ModeContext = createContext<ModeContextValue | undefined>(undefined);
+import React, { useState } from "react";
+import ModeContext, { type Mode } from "./mode";
 
 export function ModeProvider({ children }: { children: React.ReactNode }) {
   const [mode, setMode] = useState<Mode>("student");
@@ -19,12 +11,4 @@ export function ModeProvider({ children }: { children: React.ReactNode }) {
   );
 }
 
-export function useMode() {
-  const ctx = useContext(ModeContext);
-  if (!ctx) {
-    throw new Error("useMode must be used within a ModeProvider");
-  }
-  return ctx;
-}
-
-export default ModeContext;
+export default ModeProvider;
